[CHORE] : moved room related type to RoomType.tsx dquote> - removed comments dquote> - removed console.logs dquote> - cmd+s ed dquote>
