+++ frontend/components/main/room_list/Room.tsx
@@ -13,15 +13,13 @@ import {
   alert,
   lock,
   clickedLock,
-} from "@/type/type";
+  IMember,
+  ILeftMember,
+} from "@/type/RoomType";
 import { socket } from "@/app/page";
 import Alert from "@mui/material/Alert";
 import { useUser } from "@/context/UserContext";
-import { IMember } from "@/type/type";
 
-export interface ILeftMember {
-  userNickname: string, userIdx: number, imgUri: string
-}
 
 export default function Room({ room, idx }: { room: IChatRoom; idx: number }) {
   const [open, setOpen] = useState(false);
@@ -31,35 +29,34 @@ export default function Room({ room, idx }: { room: IChatRoom; idx: number }) {
   const { roomState, roomDispatch } = useRoom();
   const { userState } = useUser();
 
-    // userIdx: number | undefined;
-    // nickname: string | undefined;
-    // imgUri: string | undefined;
-    // permission?: Permission | undefined;
-
-
   useEffect(() => {
-    const ChatExitRoom = ({leftMember, owner}: {leftMember: ILeftMember[], owner: string}) => {
-      console.log("room", room)
-      console.log("owner", owner)
-      if (!leftMember){
-        roomDispatch({type: "SET_CUR_ROOM", value: null});
-        roomDispatch({type: "SET_IS_OPEN", value: false})
+    const ChatExitRoom = ({
+      leftMember,
+      owner,
+    }: {
+      leftMember: ILeftMember[];
+      owner: string;
+    }) => {
+      if (!leftMember) {
+        roomDispatch({ type: "SET_CUR_ROOM", value: null });
+        roomDispatch({ type: "SET_IS_OPEN", value: false });
         // window.alert("너 킥 당함"); // TODO : 서버에서 다섯번 보냄? 왜?
-        return ;
+        return;
       }
       const list: IMember[] = leftMember.map((mem: ILeftMember) => {
         return {
-        nickname: mem.userNickname,
-        userIdx: mem.userIdx,
-        imgUri: mem.imgUri,
-    }})
-    const newRoom: IChatRoom = {
-      owner: owner ? owner : room.owner,
-      channelIdx: roomState.currentRoom!.channelIdx,
-      mode: roomState.currentRoom!.mode,
-    }
-      roomDispatch({type: "SET_CUR_MEM", value: list})
-      roomDispatch({type: "SET_CUR_ROOM", value: newRoom})
+          nickname: mem.userNickname,
+          userIdx: mem.userIdx,
+          imgUri: mem.imgUri,
+        };
+      });
+      const newRoom: IChatRoom = {
+        owner: owner ? owner : room.owner,
+        channelIdx: roomState.currentRoom!.channelIdx,
+        mode: roomState.currentRoom!.mode,
+      };
+      roomDispatch({ type: "SET_CUR_MEM", value: list });
+      roomDispatch({ type: "SET_CUR_ROOM", value: newRoom });
     };
     socket.on("chat_room_exit", ChatExitRoom);
 
@@ -70,11 +67,10 @@ export default function Room({ room, idx }: { room: IChatRoom; idx: number }) {
 
   useEffect(() => {
     const ChatEnterNoti = (data: IChatEnterNoti) => {
-      // console.log("ChatEnterNoti ", data )
       setShowAlert(true);
       setNewMem(data.newMember);
-      roomDispatch({type: "SET_CUR_MEM", value: data.member})
-      roomDispatch({type: "SET_ADMIN_ARY", value: data.admin})
+      roomDispatch({ type: "SET_CUR_MEM", value: data.member });
+      roomDispatch({ type: "SET_ADMIN_ARY", value: data.admin });
     };
     socket.on("chat_enter_noti", ChatEnterNoti);
 
@@ -115,7 +111,6 @@ export default function Room({ room, idx }: { room: IChatRoom; idx: number }) {
     const ChatEnter = (payload: IChatEnter) => {
       roomDispatch({ type: "SET_CUR_MEM", value: payload.member });
       roomDispatch({ type: "SET_ADMIN_ARY", value: payload.admin });
-      //channelIdx 안보내줘도 될듯?
     };
     socket.on("chat_enter", ChatEnter);
 
@@ -158,7 +153,6 @@ export default function Room({ room, idx }: { room: IChatRoom; idx: number }) {
 
   useEffect(() => {
     const GoToLobby = (payload: IChatRoom[]) => {
-      console.log("GoToLobby ", payload);
       roomDispatch({ type: "SET_NON_DM_ROOMS", value: payload });
     };
     socket.on("chat_goto_lobby", GoToLobby);
@@ -176,9 +170,7 @@ export default function Room({ room, idx }: { room: IChatRoom; idx: number }) {
           channelIdx: roomState.currentRoom.channelIdx,
           userIdx: userState.userIdx,
         }),
-        (ret: number | string) => {
-          console.log("chat_goto_lobby ret : ", ret);
-        }
+        (ret: number | string) => {}
       );
     }
     roomDispatch({ type: "SET_CUR_ROOM", value: room });
@@ -187,7 +179,6 @@ export default function Room({ room, idx }: { room: IChatRoom; idx: number }) {
 
   const RoomClick = (room: IChatRoom) => {
     if (roomState.currentRoom?.channelIdx !== room.channelIdx) {
-      // TODO : 누른 버튼 색 다르게 해보기
       if (room.mode === Mode.PROTECTED) handleOpen();
       else if (room.mode === Mode.PRIVATE) {
         socket.emit(
